Finished assignment

+++ prepare/Learning03/Fraction.cs
@@ -51,6 +51,6 @@ public class Fraction
 
     public double GetDecimalValue()
     {
-        return _top / _bottom;
+        return (double)_top / (double)_bottom;
     }
 }
